Create AdvancedArithmetic.java
PROBLEM STATEMENT
Create an interface AdvancedArithmetic which contains a method signature as:
public abstract int divisorSum(int n)
Class: MyCalculator
This class implements the interface AdvancedArithmetic
Method: divisorSum()
This function just takes an integer as input and return the sum of all its divisors.
Example:
Divisors of 6 are 1, 2, 3 and 6, so divisorSum() should return 12
Sample Input: 6
Sample Output: 12
Explanation:
Divisors of 6 are 1, 2, 3 and 6. Then, 1 + 2 + 3 + 6 = 12
Note:
A class Testing is required with a main method to test the problem statement.
Use Testing class to test your solution classes and methods.